# Faster collect